+++ scripts/build_runtime_package.py
@@ -57,6 +57,7 @@ TOP_LEVEL_MODULES = {
     "a2a_tools",
     "a2a_tools_delegation",
     "a2a_tools_memory",
+    "a2a_tools_messaging",
     "a2a_tools_rbac",
     "adapter_base",
     "agent",
